First point with no compile errors. Reworking of the OBJ loader is in progress.
The stats tracker needs to be implemented, and the OBJ loading
context needs to get its objects from the results of the first pass
through the OBJ file, so that it can start filling them in the second
run. As far as I know, all objects now only consider one buffer and one
material max per part of the model. Because of the stats step, objects
also expect a total count of elements that they need to store in their
constructors, which was done to save memory.

Also did a lot of refactoring. In particular, note that the
GeometryBufferGenerator now actually generates GeometryBuffer
instances, whereas the one that used to be called
GeometryBufferGenerator is now called TemporaryVertexBuffer.